requirements.txt
streamlit>=1.20
transformers>=4.30
torch>=1.13
sentence-transformers>=2.2.2
spacy>=3.7
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl
nltk>=3.8
python-docx>=0.8.11
PyPDF2>=3.0.1
keybert>=0.7.0
scikit-learn>=1.2
pandas>=2.0
numpy>=1.23
tqdm
faiss-cpu